+++ packages/eyes-cypress/src/plugin/getErrorsAndDiffs.js
@@ -6,29 +6,28 @@ function getErrorsAndDiffs(testResultsArr) {
       if (testResults instanceof Error || testResults.error) {
         failed.push(testResults);
       } else {
-        if (testResults.getStatus() === 'Unresolved') {
-          if (testResults.getIsNew()) {
-            testResults.error = new Error(
-              `${testResults.getName()}. Please approve the new baseline at ${testResults.getUrl()}`,
-            );
-            failed.push(testResults);
-          } else {
-            diffs.push(testResults);
-          }
-        } else if (testResults.getStatus() === 'Failed') {
-          failed.push(testResults);
-        } else {
+        const testStatus = testResults.getStatus();
+        if (testStatus === 'Passed') {
           passed.push(testResults);
+        } else {
+          if (testStatus === 'Unresolved') {
+            if (testResults.getIsNew()) {
+              testResults.error = new Error(
+                `${testResults.getName()}. Please approve the new baseline at ${testResults.getUrl()}`,
+              );
+              failed.push(testResults);
+            } else {
+              diffs.push(testResults);
+            }
+          } else if (testStatus === 'Failed') {
+            failed.push(testResults);
+          }
         }
       }
 
       return {failed, diffs, passed};
     },
-    {
-      failed: [],
-      diffs: [],
-      passed: [],
-    },
+    {failed: [], diffs: [], passed: []},
   );
 }