+++ packages/eyes-cypress/test/unit/plugin/errorDigest.test.js
@@ -31,19 +31,23 @@ describe('errorDigest', () => {
         name: 'test1',
         hostDisplaySize: {width: 100, height: 200},
         url: 'url1',
+        status: 'Unresolved',
       }),
       new TestResults({
         name: 'test2',
         hostDisplaySize: {width: 300, height: 400},
         url: 'url2',
+        status: 'Unresolved',
       }),
     ];
     const passed = [
       new TestResults({
         name: 'test3',
         hostDisplaySize: {width: 1, height: 2},
+        status: 'Passed',
       }),
     ];
+
     const output = errorDigest({
       passed,
       failed,
@@ -54,12 +58,16 @@ describe('errorDigest', () => {
     // NOTE: this is a try to validate the expected output.
     // It was very hard to construct this expected string, so if this becomes a maintenance nightmare, I suggest not to try and preserve it.
     // It was mainly written for debugging purposes in order to quickly craft the output without having to run Cypress.
+
+    //  this might help:
+    // https://github.com/avajs/ava/blob/master/docs/04-snapshot-testing.md
+
     const expectedOutput = `Eyes-Cypress detected diffs or errors during execution of visual tests:
        ${chalk.green('Passed - 1 tests')}
          ${chalk.green('\u2713')} ${chalk.reset('test3 [1x2]')}
-       ${chalk.red('Diffs detected - 2 tests')}
-         ${chalk.red('\u2716')} ${chalk.reset('test1 [100x200]')}
-         ${chalk.red('\u2716')} ${chalk.reset('test2 [300x400]')}
+       ${chalk.yellow('Diffs detected - 2 tests')}
+         ${chalk.yellow('\u26A0')} ${chalk.reset('test1 [100x200]')}
+         ${chalk.yellow('\u26A0')} ${chalk.reset('test2 [300x400]')}
        ${chalk.red('Errors - 3 tests')}
          ${chalk.red('\u2716')} ${chalk.reset('test0 [4x5] : Error: bla')}
          ${chalk.red('\u2716')} ${chalk.reset('test0 [6x7] : Error: bloo')}
@@ -68,6 +76,52 @@ describe('errorDigest', () => {
        See details at: url1`;
 
     // console.log(_wrap(output)); // debugging
-    expect(output).to.equal(expectedOutput);
+    expect(output).to.deep.equal(expectedOutput);
+  });
+
+  it('should only print existing results', () => {
+    const emptyResult = new TestResults();
+    emptyResult.isEmpty = true;
+    const passed = [
+      new TestResults({
+        name: 'test3',
+        hostDisplaySize: {width: 1, height: 2},
+        status: 'Passed',
+      }),
+      emptyResult,
+    ];
+    const failed = [];
+    const diffs = [];
+    const output = errorDigest({
+      passed,
+      failed,
+      diffs,
+      logger: {log: () => {}},
+    });
+
+    const expectedOutput = `Eyes-Cypress detected diffs or errors during execution of visual tests:
+       ${chalk.green('Passed - 1 tests')}
+         ${chalk.green('\u2713')} ${chalk.reset('test3 [1x2]')}`;
+
+    expect(output).to.deep.equal(expectedOutput);
+  });
+
+  it('should handle error results', () => {
+    const failure = new Error('i failed you');
+    const passed = [];
+    const failed = [failure];
+    const diffs = [];
+    const output = errorDigest({
+      passed,
+      failed,
+      diffs,
+      logger: {log: () => {}},
+    });
+
+    const expectedOutput = `Eyes-Cypress detected diffs or errors during execution of visual tests:
+       ${chalk.red('Errors - 1 tests')}
+         ${chalk.red('\u2716')} ${chalk.reset('[Eyes test not started] : Error: i failed you')}`;
+
+    expect(output).to.deep.equal(expectedOutput);
   });
 });