+++ packages/eyes-cypress/CHANGELOG.md
@@ -2,9 +2,10 @@
 
 ## Unreleased
 
+- fix output with empty tests
 - using new concurrency model
-- updated to @applitools/eyes-sdk-core@12.6.1 (from 12.5.7)
-- updated to @applitools/visual-grid-client@15.2.0 (from 15.1.1)
+- updated to @applitools/eyes-sdk-core@12.6.2 (from 12.6.1)
+- updated to @applitools/visual-grid-client@15.2.1 (from 15.2.0)
 
 ## 3.15.5 - 2020/11/10